% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_label_point.r
\name{find_best_candidate_points}
\alias{find_best_candidate_points}
\title{select the best point from candidate points for all polygon
best candidate point: has the largest min_dist to polygon
candidate points: center, mid_x, mid_y, mid_xy, mid_yx}
\usage{
find_best_candidate_points(current_id, label_centers, max_values, mid_dist_df)
}
\description{
select the best point from candidate points for all polygon
best candidate point: has the largest min_dist to polygon
candidate points: center, mid_x, mid_y, mid_xy, mid_yx
}

