% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_label_point.r
\name{find_mid_points}
\alias{find_mid_points}
\title{find point in polygon which has min y-distance with y of center
find_mid_points()}
\usage{
find_mid_points(current_id, label_centers, polygon_data)
}
\description{
find point in polygon which has min y-distance with y of center
find_mid_points()
}

