% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_label_point.r
\name{mid_range}
\alias{mid_range}
\title{calculate center (not centroid) of each OT (ggplot2-book.pdf p.80)
centers are used for putting labels in plots (nrow = 63)}
\usage{
mid_range(x)
}
\description{
calculate center (not centroid) of each OT (ggplot2-book.pdf p.80)
centers are used for putting labels in plots (nrow = 63)
}

