% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_basic_fill.r
\name{plot_basic_fill}
\alias{plot_basic_fill}
\title{Plot map with labels/graphics inside each polygon}
\usage{
plot_basic_fill(map_data, label_x = ot_label_points$best_x,
  label_y = ot_label_points$best_y, label_text = ot_label_points$id,
  fill_data = NULL, pal = "Greens", title = NULL, label_size = 4,
  family = "Helvetica", alpha = 1, ...)
}
\arguments{
\item{map_data}{polygon data for plotting map}

\item{label_x}{longtitude positions for labels/graphics}

\item{label_y}{latitude positions for labels/graphics}

\item{label_text}{text or graphics of the labels}

\item{fill_data}{OPTIONAL: a column in map_data for filling polygons with colours (alwasys merge with map_data)}

\item{pal}{colouring palette, as in ggplot2}

\item{title}{title of the plot}

\item{label_size}{text size of the labels}

\item{family}{font of the labels}
}
\value{
a ggplot2 plot
}
\description{
Plots map and places labels/graphics inside each polygon.
Colouring the polygons with extra data is also possible.
}
\examples{
see "use-cases/usecase-plot_basic_fill"
}

