% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_categorical.r
\name{plot_categorical}
\alias{plot_categorical}
\title{Plots pie charts for each region in a map}
\usage{
plot_categorical(map = "Ortsteil", plotType = "bar", inputDf = bmiDf,
  filterValue = 0, colList = NULL, legendText = NULL, textSize = 0.8,
  textFont = "Helvetica", legendPos_lx = 12.35, legendPos_ly = 51.24,
  legendPos_rx = 12.357, legendPos_ry = 51.27, pieRadius = 0.0045,
  sizeX = 0.3, sizeY = 0.3)
}
\arguments{
\item{map}{polygon data of the map}

\item{plotType}{"pie" for pie chart and "bar" for bar chart}

\item{inputDf}{categorical data to be plotted, three columns with colnames=("id", "category", "num"). "id": region id}

\item{filterValue}{sets filter for in each region. If sum(num) of a region is smaller than filterValue, no chart is displayed}

\item{colList}{color list for pie chart. Default list is inside the function}

\item{legendText}{a vector of legend text for each category. Default is values in column "category"}

\item{textSize}{textSize for legendText}

\item{textFont}{font of the legendText}

\item{legendPos_lx,}{_ly, _rx, _ry: lower left and upper right coordinates of the color legend}

\item{pieRadius}{set radius for pie chart}

\item{sizeX}{length of x axis for bar chart}

\item{sizeY}{length of y axis for bar chart}
}
\value{
map of Leipzig with pie charts in each region
}
\description{
Plots categorical data as pie charts for each region of Leipzig.
}
\details{
User can define which labels are placed in which of the three label regions.
}
\examples{
see "use-cases/usecase-category_chart_bmi.r"
}

